[ATC] Improve C abstract syntax and fix C dynamic semantics.
Given that the body of a function definition must be a compound statement, we
build this requirement into the abstract syntax by using a list of block items
(implicitly forming a compound statement) instead of a general statement.

This also makes the static semantics slightly simpler, as it does not need to
check that the body statement of a function definition is a compound one.

Adapting the dynamic semantics to this change also fixes a bug related to the
one fixed in a recent commit for the static semantics. Namely, that the
execution of the body of a function must not start a new scope, but must instead
happen in the scope that contains the function's arguments.