[Yul] Revise definition of the static semantics.
Avoid the flags saying whether we are inside functions and loop initialization
blocks and loop body blocks, which were used to check proper occurrences of the
leave, break, and continue statements. Instead, have the checking functions
calculate sets of possible termination modes, which are used to enforce
restrictions. This approach should also make it easier to prove static
soundness.