[Yul] Fix bug in dynamic semantics.
Before this commit, the function environment was a finite map from function
names to function information (i.e. inputs, outputs, body), that was extended as
block were entered, also across function calls. However, when calling a
function, it is necessary to reduce the functions in the function environment to
the functions in scope for the called function.

Thus, the function environment is now a stack (i.e. list) of finite maps
(i.e. scopes), that is suitably reduced when calling functions. A function scope
is pushed when entering a block, and popped of zero or more scopes when calling
a function. Function environments are just recursively passed to the execution
functions; they are no longer threaded through, and so there is no need to pop
anything when exiting a block (we simply go back to the function environment
before the block.

At the same time, we have simplified the collection of the functions that form
the function scope to push onto the function environmetn stack, by first
filtering the function definitions out of the statement list in a block, and
then turning the list of function definitions into a function scope. The
resulting code is more orthogonal and clear.