[ATC] Improve handwritten testing functions.
Have their main() return 0, so it considered a success in a shell script.
Without this, the run.sh script stops after executing the first binary, because
apparently the run is treated as a failure, and the set -e in the script stops
things at the first error.

With these additions, all the binaries are executed.