[ATJ] Move the :NO-AIJ-TYPES check to pre-translation.
Performing these checks during the initial input processing was causing "false
positives", in the sense of rejecting ACL2 code that would have resulted in Java
code that does not use AIJ types. This is because some pre-translation steps
remove some ACL2 code that is not translated to Java, so the checks should not
apply to that code. This commit "moves" the checks into a new pre-translation
step, which occurs after the code to be removed is actually removed. Overall,
the code that performs these checks has become simpler.

It may be even better to move these checks to a later pre-translation check,
after all the type annotations have been added by another pre-translation checks
(which currently occurs after this new one): this way, it may suffice to check
all the types, without worrying about specifiy terms, functions, etc. This is
something for future consideration.