Added ld-result.  Made progress on verifying termination and guards for remove-guard-holders.
Quoting :doc note-8-5:

  The undocumented last-ld-result feature has been replaced by a new
  documented feature, a [ld-result] that records command input/output
  history.  Thanks to Eric Smith for requesting this feature.

Also, at Eric Smith's request I did some work on verifying termination
and guards for the built-in function, remove-guard-holders.  (That
work had already been completed for remove-guard-holders-weak, an
older version of remove-guard-holders.)  The new file (not a book),
books/system/remove-guard-holders-strong.lsp, contains the work so far
and information about completing it.  Anyone interested in completing
this work should contact me, as noted near the top of that file.