[apt] Improve def-equality-transformation.
Now all of the transformations it creates have a :normalize option (t by default).  Using a value of nil will hardly ever be necessary, but it does seem necessary a few calls of expand-lets (e.g., in a complicated derivation to keep ACL2 from going out to lunch on one of the proofs).