[ATJ] Add a post-translation step.
As explained in the added documentation, this Java-to-Java transformation turns
each nullary method whose body consists of a single return statement whose
expression does not call other methods into a static final field (i.e. "Java
constant") initialized with that expression and the method modified to return
that field instead of the expression. It is expected that JIT may inline calls
of the so-transformed method.

This is a bit like generating Java constants (i.e. static final fields) from
ACL2 nullary functions, but in a more "uniform" way, in the sense that we also
generate nullary methods.

Thanks for Stephen Westfold for working on an example of Java code generation,
and for providing some ideas, that eventually led to this ATJ extension.