Fixed some bugs in processing DO loop$ expressions.
Thanks to J for pointing out a bug, which led me to find some others.
A key user-visible change is that side effects with cw in DO loop$
expressions are now visible even when code is not guard-verified.  See
for example tests added near the end of
books/projects/apply/loop-tests.lisp.