[X86ISA] Explicate some macro arguments.
This is as discussed with Shilpi Goel and Eric Smith.

A few macros related to floating-point numbers had implicit arguments, in the
sense that they had no arguments but expanded to terms that assumed that certain
variables were bound in context. These have been turned into explicit arguments,
adapting the callers.

Thanks to Eric Smith for proposing this change.