Improved prove$, simplifying the interface by simplifying error and output handling, in particular eliminating funky option :with-translate-error; also added option :skip-proofs.  Deprecated recently-added utility prover-steps-counted, since it is the same as last-prover-steps.
See :doc prove$, in particular for explanation of the :skip-proofs
option.  Thanks to Eric Smith for discussions leading to the other
improvements to prove$.

Changed :doc prover-steps-counted to say that it's decremented and
point to :doc last-prover-steps.  Improved :doc last-prover-steps by
folding in some of the documentation that was in :doc
prover-steps-counted.