[ATC] Rework DEFSTRUCT.
Now this macro generates a recognizer for an appropriate subset of the type of
all structures that is part of the deep embedding. It no longer generates a
FTY::DEFPROD. The subset relation facilitates the generation of proofs by ATC,
obviating the need for conversions.

The macro also generates an explicit fixer and fixtype.

The generated readers and writers are defined in terms of the generic ones that
are part of the deep embedding, but they have tighter types and they never
return errors. To prove this tighter typing, the macro also has to generate
theorems, proved by leveraging some support theorems that have been added as
part of the implementation of this macro.