[ATC] Add support for some structure theorems.
This is for the result theorems of functions involving structs.

DEFSTRUCT has been extended to record return type theorems for readers and
writers.

Proof generation has been extended to include, in the proof hints for the result
theorems, the return type theorems of the readers and writers of the structs
that precede the function in the list of targets passed to ATC.