[ATC] Extend mapping of types to recognizers.
Include structure types. This requires a dependency on the ACL2 world, because
we need to retrieve the recognizer from the DEFSTRUCT table. (A previous commit
extended the table to store the recognizer.)

This is letting more of the generated result theorems about the C code
succeed. More extensions to do to get all the current examples working.