Make it easy to build saved_acl2 with a specified optimization level for SPACE.
For example, to get optimization with (SPACE 1) instead of the default
of (SPACE 0), use ACL2_SPACE=1 in the "make" command that builds
saved_acl2.  (Actually 1 is already the default for CMUCL.)

I may change the default DECLAIM form for optimizations to specify
(SPACE 1) instead of (SPACE 0), since I have an example where this
works much better for a specific application with SBCL.  I plan to add
a release note about SPACE optimization after deciding about that.