[make-flag] Improve defthm-flag.
When defthm-flag is called with no arguments, it creates a template form for the user to fill out.  That form includes placeholder hyps for each template theorem.  Previously these were always literally (AND HYP1 HYP2), with the user expected to replace the meaningless HYP1 and HYP2 with meaningful hyps (or delete them).  Now, for a function with a non-trivial guard, that guard is used for the hyp instead, since it is fairly likely to be needed in theorems about the function.  The user is still welcome to edit the hyps, of course.