[ATC] Update model (and "API") of arrays.
This is in preparation to an upcoming extension of the model of values that will
include arrays. We want the model of arrays in the shallow embedding to be a
subset of the model of values: this commit changes the model so that it will be
a subset.

The "API" changes slightly as follows:

<TYPE>-ARRAY => <TYPE>-ARRAY-OF

i.e. the constructor has a new name now. This is because <TYPE>-ARRAY now take
two arguments, due to the change in internal structure, and so a new
one-argument constructor, <TYPE>-ARRAY-OF, has been introduced as a substitute.