Fixed soundness bug in the use of SATISFIES in stobj field type-specs.
Quoting :DOC note-8-5:

  A soundness bug in [defstobj] has been fixed that allowed a field to
  contain the [type-spec] (SATISFIES pred) when pred has a [guard]
  other than t, which generated an unsound guard theorem.  Thanks to
  Eric Smith for sending the following example.

    (defstobj s (field :type (satisfies evenp) :initially 4))

    (defthm bad
      nil
      :rule-classes nil
      :hints (("Goal" :use (:instance (:guard-theorem fieldp) (x t)))))

Interestingly, it turns out that ACL2 almost already avoided that bug.
For example, if you were to define my-evenp exactly like evenp and
then use my-evenp above in place of evenp, then you'd get an error,
even before the bug fix.  The problem was that the check was done with
function chk-unrestricted-guards-for-user-fns, which -- as the name
implies -- only checked user-defined functions.  My fix was simply to
check that all unary functions in the translated type-spec, whether
user-defined or not, have a guard of t.

Also, updated :DOC defstobj to explain the SATISFIES restriction.