[ATC] [C] Refactor the abstract syntax.
These is no change in the definition of the abstract syntax, but its definition
is now under the C language formalization rather than ATC, while the
abstrat-syntax.lisp file includes it from the language formalization and adds
some utilities that are currently only relevant to ATC.

This is part of moving the current formalization of C used by ATC from ATC to
the C language formalization proper (i.e. the deep embedding of C in ACL2). Even
though the formalization only covers a subset of C, it belongs to that
directory; and it can be extended over time.