Fix "make basic" to avoid quicklisp.  Add a bit to :DOC books-certification about need for OpenSSL for quicklisp.
Quoting :DOC note-8-5-books (the first line is a header):

  Default Build Target, `basic', Avoids Quicklisp Dependency

  The default `make' target for certifying the [community-books], which
  is still `basic', now certifies slightly fewer books so that it
  does not depend on Quicklisp.

Below is an email I sent on 6/8/2022 to the acl2-books list.  I didn't
see a reply, so I've taken care of it.  Anyone who objects to these
changes can of course put forward further changes.

..... email 6/8/2022 to acl2-books

Hi acl2-books,

I've received a suggestion that the ACL2 installation page
[https://www.cs.utexas.edu/users/moore/acl2/v8-4/HTML/installation/installation.html]
should be updated so that Step 4, which suggests "make basic", points
to :DOC ccl-installation-mac-brief, in case users run across a
quicklisp error related to openssl.  I really hate to further
complicate the installation instructions in that way.  I also think
it's weird for a small target like "basic" to bring in trust-tag
external tools like quicklisp.  It seems to J and me that "basic"
should just test book certification (without failing!) and to give
people a bit more of a proof environment than comes with the system.

Is anyone willing to volunteer to change "basic" to avoid quicklisp?
This might require moving xdoc out of some basic books.  Even if basic
were just an arithmetic book, J and I would be much happier than we
are now.

All that said, I also think that since the installation instructions
point to :DOC books-certification, that topic should either mention
installation of openssl or point to :DOC ccl-installation-mac-brief.
I'd be happy for someone who understands this issue to make that
change; otherwise I probably will.

Thanks,
Matt