[ATC] Generalize some code.
This is for the generation of execution theorems for structure readers and
writers. Since we'll need to generate multiple ones for each array members
(unlike a scalar member, which needs just one for reading and one for writing),
the code has been generalized to produce lists of events and theorem names.