Fixed a bug in handling of local commands.  Tweaked an error message in logical-name-type.
The latter is thanks to a discussion with Alessandro Coglio.  Below
I'll elaborate the former, a bug in handling of local commands.

That bug was introduced after Version 8.4 -- probably in commit
b1413911423bfb7c1266eeefdf83abe5c59a798b -- pertains to support for
local portcullis commands.  Thanks to Alessandro Coglio for sending an
example (where :a! in a break-rewrite loop caused a translate error)
that helped me debug the problem.  Here is a very simple example of
the problem; this formerly caused an error, but doesn't now.

(ld '((local (ld '(3)))))

The problem was that the command '(3) was replaced by (LOCAL '(3))
before evaluation.  Now, that replacement happens after evaluation,
and only for a command that evaluates to an error triple, as explained
further below.

The following has been *deleted* from :DOC note-8-5 -- it now seems
too low-level, although it's still approximately true (evaluation
though is unaffected; the only change is in the command that is
stored).  Further explanation is below.

  Suppose [ld] is called in the scope of [local], in particular, as
  with (local (ld ...)).  Then for each [command] C read by that call
  of ld that is not already of the form (local ...), C is read as
  though it had been (local C).  (This change has been made in
  support of [local] [portcullis] [events], a new feature described
  further below.)

The following, which this commit adds to :DOC ld-history, explains
more precisely how ld-history is affected by LOCAL.

    * When an entry is saved for a command C that is evaluated in the
      context of a call of [local], where C evaluates to an
      [error-triple], then C is stored in the entry as (local C)
      unless C is already a call of local.  (Technical Aside: This
      behavior supports local [portcullis] commands.)

Also updated a related paragraph already in :DOC ld.