[ATC] Separate and extend user doc of DEFSTRUCT.
Now the user documentation of DEFSTRUCT is in a separate documentation file, and
it has been significantly extended and improved to be more like the user
documentation of DEFOBJECT and of many other tools. In particular, now the user
documentation of DEFSTRUCT explains in more detail the generated functions and
fixtypes.