[ATC] Improve the C dynamic semantics.
Reformulate the execution of casts in terms of the much simpler integer
conversion functions, instead of via a lot of different cases and using the
conversions from the shallow embedding. Adapt the proofs of the symbolic
execution rules for executing casts, which before had very simple proofs, with
more elaborate proof hints needed to bridge the just (intentionally) created gap
between the deep and shallow embedding definitions.