[C] [ATC] Simplify and move definition of PROMOTE-VALUE.
This is now part of the integer operations in the deep embedding, where it
should be. The definition is simpler and clearer, in terms of the operation to
promote types.

Adapt ATC's symbolic execution to the change. Prove the old definition as a
theorem, and use it in the existing proofs, for minimal change. This may be
removed eventually, as the proofs are evolved to use the new definition (and
more elaborate hints, to bridge the now slightly larger gap).