[C] Strengthen two guards.
These are the guards of PROMOTE-TYPE and PROMOTE-VALUE. Now they operate on
arithmetic types and values, instead of any types and values. This restriction
makes sense as these functions should only be used on arithmetic types and
values.

Also add theorem saying that PROMOTE-TYPE returns a promoted arithmetic type, as
expected of course.