[C] Rework some code.
Have VALUE-INTEGER always return a value (never an error), with a stronger guard
requiring the ACL2 integer to fit in the type. This makes more sense, as it is
similar to the individual constructors for the different integer values.

Move the range checking logic to CONVERT-INTEGER-VALUE, which is more in sync
with the ISO standard's description.

Adapt and add some theorems.

Adapt some proofs, e.g. for the symbolic execution rules for casts.