[ACL2s] Updates to interface functions supporting ACL2s systems programming. Updated itest? and itest?-fn as follows.
1. What the forms return has been changed to allow one to easily determine
whether the conjecture was proved or not.

2. The forms now also return witnesses, if any were found.

3. All counterexamples and witnesses found are returned, not just the number requested.
This is a better way of doing things for interface functions: the test? versions
return no more than the number of counterexamples and witnesses the user requested
to avoid cognitive overload, but the itest? forms are being use programmatically and
in this context, it is better to have access to all the counterexamples found.

4. Documentation is provided in the comments.

5. Added a simple version of itest?-fn that doesn't take hints override stuff.

6. Updates to extra.lsp to make it consistent with the updates.