[ATC] Change the struct pointer type designation.
This breaks compatibility, but in a way that is easy to fix.

Before this commit, guard conjuncts (STRUCT-<TAG>-P X) signified that the type
of the formal X was pointer to struct TAG. This commit now requires that
conjunct to be (C::POINTER (STRUCT-<TAG>-P X)). The motivation is that soon
(STRUCT-<TAG>-P X) will signify that X is a struct, i.e. passed by value, which
is an upcoming capability in ATC. Also in the future, it should possible to use
the C:POINTER wrapper (which is just identity, similarly to the C::CONDEXPR,
C::DECLAR, and C:ASSIGN wrapper already in use) to designate pointer types,
e.g. (C::POINTER (C::SINTP X)) could means that X is a pointer to an integer.