Fix testing failure with whitespace-only change.
There was no failure when I ran "make" with target
"regression-everything-fresh", before committing each of my recent
changes to the ACL2 system.  I believe I've seen testing failures (for
runs that didn't clean the books first) like this before:
(verify-termination project-dir-lookup) expands to a make-event that
caused a defun to be stored in the book's certificate, but the defun
had changed.