Updated :DOC xdoc::save and :DOC io.
For :DOC xdoc::save, emphasized that saving manual to the current
directory is a bad idea.  Thanks to Warren Hunt for the suggestion.

For :DOC io, clarified that the file-name arguments are interpreted by
Lisp relative to the directory where ACL2 was invoked.  Thanks to Eric
Smith, Eric McCarthy, and Alessandro Coglio for a discussion leading
to this change.