Changed project-dir-alist to be a world global (formerly a state global).  Added er-hard? and project-dir-alist to *acl2-exports*.
Here is a rather technical explanation.

The change for project-dir-alist allows resolution of so-called
sysfiles, (:dir . "dir"), to full pathnames using just the world, not
the rest of state.  I did this to improve support for moving project
directories, with better include-book redundancy when using an
executable produced by save-exec.  This is partly why the notion of
full-book-name is likely to change soon, to allow a full-book-name to
be a sysfile (in general, not merely in .cert files), something J and
I agree would be a good idea.  (I've started working on that.)