[APT] Improve RESTRICT's proof generation.
Extract the limited-simplified guard obligations for the restriction terms, and
generate a VERIFY-GUARDS event that does limited guard simplification. This
makes the generated guard proofs more robust: we ran into a case in which ACL2
was simplifying away a guard hypothesis, making the application of the guard
theorem of the old function (which included that hypothesis) fail.

Thanks to Eric Smith for running into the issue in an example, and for useful
discussions.