[ATC] Speed up ATC certification.
Specifically, speed up the generation of the symbolic execution rules for C
binary operators. More optimizations should be possible, there and elsewhere,
but these had become significantly slower after some adaptations of their proofs
to the reformulation of the definition of some binary operations in sole terms
of the deep embedding.