[ATC] Extend support for structs.
It is now possible to represent assignments to scalar members of structs by
value, generating assignments of the form s.m = e. This involves generating
additional struct-specific symbolic execution rules, and extending the code that
generates statements to recognize and translate the pattern. Proof generation
for this new feature is supported.

Support for doing the same with array members of structs is forthcoming.