[C] Extend static semantics.
Check that element types of array types are complete. This involves a
restructuring of the ACL2 functions that formalize these checks, which no longer
check type specifier sequence and abstract object declarator independently, but
rather check the first and then the second in the context of the first,
augmenting the type as the declarator's layers are recursively processed.  Also
change these functions to return the denoted type in case of success.