[ABNF] Simplify formulation of concrete syntax.
Instead of using the plugging operation, just use APPEND with the list of needed
core rules. This avoids the dependency of the definition of the concrete syntax
of ABNF on the pluggin operations, which in some sense they should build on the
formal definition of the ABNF notation, which should be defined independently
from them (as it is now).