[C] Extend abstract syntax.
Add a notion of file, which is more general than translation unit, and is meant
to capture the content of files, with constructs both before and after
preprocessing. This will be later generalized to a notion of file sets, i.e.
collections of related files that form (parts of) programs.