[ATC] Improve proof generation.
Stage the proof of correctness of C functions in two stages: (1) symbolic
execution of the C function while keeping the ACL2 function closed; (2)
completion of the proof by opening the ACL2 function.

The motivation is to reduce the performance impact of case splits during the
proof. During the first phase, the definition of the ACL2 function causes no
case splits. It looks like the symbolic execution still leads to multiple goals
corresponding to the case splits in the C function, which then get inevitably
combined with the case splits of the ACL2 function, presumably leading to a
quadratic number of subgoals; however, this approach may still reduce some
transitory case splits during the symbolic execution.

This seems to have a mild effect on the tests under
`[books]/kestrel/c/atc/tests/`, comparing their run before and after this
change. But it was just a possibly noisy measurement.

Thanks to Eric Smith for useful discussion on this (a while ago).