[ATC] Fix bug in handling of LETs.
We need to pass the updated variable-term alist to the body of an assignment
LET, in the same way as the other kinds of LETs (and MV-LETs), in order to
compute the correct instantiations for the guard theorems and perhaps other
theorems. So far none of our tests ever exposed this issue, which would have
manifested in a failed generated proof.