Fixed bug in the tool, include-book-paths.
This bug was introduced by recent work supporting relocation of book
directories.  Thanks to Eric Smith for pointing out the bug with an
example.

(Aside: The bug was a switch in binding the first two values returned
by parse-book-name.  Sheesh.  I've checked that all the other changes
made for parse-book-name (to support relocation) are correct.)