Move several projects from RTL to ACL2 package.
Move the following projects to the ACL2 package:

books/projects/quadratic-reciprocity
books/projects/curve25519
books/projects/shnf
books/workshops/2022/russinoff-groups

The rationale is that they are not about RTL and so should not be in that package.

This change means that some function and theorem names, including primep and divides, are now in the ACL2 package.

Also renamed the following (to avoid name clashes with pre-existing names in the ACL2 package):
perm -> permutationp
sublistp -> sub-listp
sum$ -> sums$