[ATC] Generate additional theorems.
When calculating the typed formals of a function, also generate theorems, one
per form, saying that each formal has certain properties that are needed for
upcoming proofs. The names of these theorems are stored in the variable
information for the formals, which had already been extended with a slot for
that purpose, in a previous commit.