[C] Finish moving execution to deep embedding.
Now the whole C dynamic semantics is under the `language/` subdirectory, which
contains the formalization of (our subset of) C in ACL2, i.e. the deep embedding
of C in ACL2.

The `atc/` subdirectory contains a mix of the shallow embedding and the code
generator. We plan to separate out the shallow embedding into a new
subdirectory, because it is more general than the code generator (in particular,
it can be used for code lifting).