[C] Add notion of file set.
This notion is that of a collection of related files, which form a program, or a
library, or some other program fragment. Currently this is just a wrapper for s
single file, but it provides the appropriate structure for extending this to
multiple files. In particular, the name of the file is part of this data
structure.

Besides adding a fixtype for file sets to the abstract syntax, the static
semantics has been extended accordingly. Preprocessing now operates on file sets
(still essentially a no-op for now, but it has the appropriate target of
operation), yielding a single translation unit for now.