[Std/util] Put some code in logic mode.
Avoid the `(program)` declaration in the middle of the file, and instead add
`:mode :program` to the functions that are not immediately accepted in logic
mode. Several functions, previously in program mode, are now accepted in logic
mode, with the only addition of a `:guard t` to cause their guard verification.
It may be actually easy to put more functions in this file in logic mode (and
guard-verified), but this is just a simple change for now.

Also clean up a few commented-out fragments that do not seem to serve a purpose
at this point.