[ATC] Fix bug in `defobject`.
For the case in which the object has no explicit initializer, we generate an
initializer funtion that sets every element to 0. However, for some reason the
code was using the `<type>-dec-const` constructors for that purpose, and was
attempting to generate non-existing constructors like `uchar-dec-const`, instead
of using `sint-dec-const` and a cast in this case. The issue was actually fixed
by using the fixtype constructors for the type, which are uniform across the
integer types: here we just need to specify the value, it does not have to
represent a C expression.

Thanks to Eric Smith for running `defobject` on an example that exposed the
issue.