[ATC] Refactor an ATC input.
The single `:output-file` input has been replaced by two separate inputs,
`:output-dir` and `:file-name`. The first one is the directory where the
generated file is written, and the second one is the name of the file without
the `.c` extension. This is in preparation for generating multiple files: it
makese sense to specify the directory where these files go, separately from the
names of the files. The reason for removing the `.c` extension is that we will
soon generate also a `.h` file, so the `:file-name` input specifies the common
part of the `.c` and `.h` file. (For future extensions in which we may generate
more than these two files, we will also need a more general way to specify all
the file names, but it seems that the `:output-dir` input should be "stable".)

At the same time, we added restrictions on the characters that may appear in
`:file-name`. They may be only letters, digits, underscores, and dashes.  This
way we easily exclude the presence of directory paths (slashes) and file
extensions (with dots). It is easy to allow more characters if needed, but it
seems that the current one may be generally sufficient.

Adjust all the calls in the test.