[ATC] Fix bug in proof generation.
Besides the type prescription rules, saying that the measure is a `natp`, for
the directly called functions, we also need those rules for the indirectly
called functions. The reason is that the measure of an indirectly called
function may be referenced in the limit bound hypothesis, and the symbolic
execution needs to know that is it a non-negative integer in order to discharge
the hypothesis that the limit is not 0.

Thanks to Jim McDonald for running ATC on a test that exposed this issue.