[ATC] Improve tests.
Now all of the tests also generate a header file, except for one that tests
external objects: it turns out that the current treatment of external objects,
with respect to what goes into the header vs. source file, is inadequate, and
needs to be refined. This will be done soon, and the remaining test will be
modified to also generate the header.