[ABNF] Add new `deftreeops` tool.
Currently this contains the functionality in `defgrammar` to generate
grammar-specific tree matching predicates. It makes sense for this functionality
to be a separate tool because a grammar may not always defined via `defgrammar`,
but by applying grammar-building operations (e.g. plugging, or simply appending
rules) to grammars (which are defined via `defgrammar`, i.e. from files). That
functionality of `defgrammar` will be removed from `defgrammar`, since it is now
available in this new tool.

We plan to extend this new tool to generate additional grammar-specific tree
operations (which justifies its more general name). In particular, high-level
accessors of trees that match components of the grammar.