[C] Improve static semantics.
Refine the treatment of object declarations. Introduce an explicit notion of
definition status of an object (i.e. variable), as undefined or tentatively
defined or defined, according to the C standard. Elaborate the operation to
extend a variable table to properly handle all 10 cases of definition statuses
in table and declaration, including the case of an absent variable in the table.