[ATC] Improve generation of external objects.
Use `extern` for declarations in headers, and always generate a declaration,
with initializer if present, and without `extern` in the source file. This
ensures that there is exactly one definition, or tentative definition, in the
source file, and not in the headers, which may otherwise cause compilation
failures if the header is included by an external (i.e. non-ATC-generated)
source file.