[C shallow] Improve formulation of the integers.
Instead of using `fty::defprod`, which forces a particular representation with
particular operation names, use more manually and explicitly defined fixtypes
for the integer values in the shallow embedding. The current re-formulation is
actually almost completely equivalent to the one with `fty::defprod`, but it
opens the door to future planned changes. In particular, we may use better
function names for the accessors. Furtheremore, as the model of values in the C
deep embedding may evolve to be more general, we will still have the ability to
adapt the exact representation of integers in the shallow embedding to the new
model, leaving the API unchanged.