[ATC] Extend modular proof generation.
Do not stop at `mbt` and `mbt$` tests, because those tests are implied by the
guard, so we can still geenrate reliable `okp` lemmas in the 'then'
branches. However, stop modular proof generation after the 'then' branches,
since we are not yet generating the modular theorems for the conditionals
themselves.