Fix ACL2(p) build bug.  Eliminated *hcomp-loop$-alist*.
Thanks to Eric Smith for pointing out the ACL2(p) build bug.

For ACL2(p), ran "make basic" with ACL2_CUSTOMIZATION set to
acl2-customization-files/parallel-resource-based.lisp.  For ACL2, ran
the usual regression-everything.

The rest of this message explains elimination of *hcomp-loop$-alist*,
but I don't expect anyone to read it (unless a future ACL2 developer
or I have reason to).

....

I've determined that the Lisp special variable *hcomp-loop$-alist* is
not necessary.  Its only use was to help the loop$ macro find its
translation during early loads of compiled files; loop$ was happy to
find that translation with (loop$-alist-term loop$-form
*hcomp-loop$-alist*).  However, we already prohibit the use of loop$,
indeed of anything depending on apply$, in defconst, defpkg, and
defmacro forms, as noted in the following sentence from an error
message.

  We prohibit certain events, including DEFCONST, DEFPKG, and
  DEFMACRO, from being ancestrally dependent on loop$ and lambda$
  expressions.

So just as we have no need to record translations of lambda$ for early
loads of compiled files, we have no such need for loop$, either.  That
is: we don't do anything analogous for 'lambda$-alist (there's no
*hcomp-lambda$-alist*), so if there's a reason not to eliminate
*hcomp-loop$-alist*, then presumably there's already a problem with
our handling of lambdas or lambda$ forms.

More evidence that we don't need *hcomp-loop$-alist* is this comment
in simple-translate-and-eval.

; We believe (as of 10/19/2021) that the point of checking for lambdas here is
; to avoid the need to consult world global 'lambda$-alist or 'loop$-alist when
; doing an early load of compiled files....