[ATC] Improve modular proof generation.
For boolean expressions, also generate a conjunct saying that the term satisfies
`valuep`, since it facilitates the use of the theorem in subsequent
theorems. (As an alternative, we may instead track C types for boolean
expressions at some point, obviating the need for this extra conjunct, and
adding a rule to the hints of the downstream theorem.)

Also make the proof more robust, in case the term or its negation is in scope
and gets rewritten to `t` or `nil`.