[ATC] Extend modular proof generation.
Also generate theorems for block items and singleton lists of block items for
conditional statements. These are temporarily not actually submitted because
they fail in some examples, due to ACL2 case splitting even though the supplied
rules should be enough to prove the theorems without splits. We plan to address
this soon.