[ATC] Extend the modular proof generation code.
Add code to generate theorems that are like the guard theorems but with `if`
replaced with `if*`. These events are not being submitted yet, because the
initial hints caused at least one example to run slow due to `if` splits in a
moderately large guard theorem; a second attempt at hints (the current ones in
the code) makes that example work quickly but is causing some other failures,
presumably due to rewritings of conditions to `nil` or `t` that prevent the
final `:by` hint from working. A more elaborate proof strategy for these
theorems is being investigated.