[ATC] Extend `defstruct`.
For each array member, generate a writer for the whole array. This has no
counterpart in the C code, but it is useful for external tools.

Thanks to Eric Smith for suggesing the addition of these writer, as well as of
the whole-array readers and of the struct constructors, which were accomplished
in previous commits.