[ATC] Enable modular proof generation for `||`.
Thanks to Matt Kaufmann for providing a variant of an ad hoc rule that I tried
to overcome an introduction of a lambda in a `hide` but didn't work. His rule
suffices to make all our current uses of ATC work, so for now there is no need
to switch from `hide` to a custom wrapper.