[ATC] Refactor and fix some code.
Put the code to generate code for `(sint-from-boolean ...)` terms into a
separate ACL2 function. In the process, also add a missing validation check,
whose need was uncovered by looking into the finer-grained modular proofs.  The
issue is that when `sint-from-boolean` is applied to a boolean term, we must
ensure that the expression generated from the bolean term has type int (not
other integer types). Without this check, one can generate C code that is not
statically well-formed (the well-formedness theorem fails).