[ATC] Complete bug fix from previous commit.
The validation condition must be stronger in fact: not just that the argument of
the `sint-from-boolean` conversion generates an expression of C type `int`, but
that the argument in question is an `and` or `or`, and not a conversion
`boolean-from-<type>`, or specifically `boolean-from-sint`. The reason is that
these conversions, which are not represented in the C code, are not inverses:
`sint-from-boolean` always returns 0 or 1, while `boolean-from-sint` coerces any
non-zero value to `t`. A simple example with `(sint-from-boolean
(boolean-from-sint x))` generates a correctness theorem that doesn't prove.