[ATC] Improve representation and generation of code.
Now array parameters have an array type, instead of an (adjustment-equivalent)
pointer type. In generated code, this makes it more clear that this is meant to
be an array, as opposed to a pointer to a single integer -- of course there is
no real distinction in a C implementation, but conceptually there is meta data
associated with C objects, which distinguish the two cases in the formal
semntics.

In addition, for ATC's own implementation, this makes the distinction more
apparent between arrays and pointed-to integers, which will be supported in
upcoming extensions.

This required a number of adaptations to several parts of the ATC
implementation.

The user documentation has been updated to reflect the slightly different
generated C code.