[C shallow] Rename accessors of C integers.
The functions `<type>->get` (i.e. `uchar->get` etc.) have been renamed to
`integer-from-<type>` (i.e. `integer-from-uchar`), which should make it clearer
that we are obtaining an ACL2 intger from a C integer; in fact, these accessors
are like conversions.

We also plan to rename the constructors, i.e. the inverses of the accessors,
according to a similar pattern. We will do that in a separate commit, since
these renamings involve a lot of changes, beyond the community books.

Thanks to Eric Smith for contributing to the choice of new names.