[ATC] Allow `mbt` in guard conjuncts.
In the guards of the ATC target functions, we now allow `mbt` to wrap conjuncts
that determine the C types of function parameters. This is useful when there are
stronger guards that imply the guards for the types: they can be added
redundantly (possibly via some automated transformation), making it easy for ATC
to determine the types, while making the redundancy provably explicit.