[built-ins] Add another disabling option.
Define a constant with the built-in rewrite rules that rewrite generic calls of
certain functions to default values when hypotheses are satisfied saying that
the arguments are outside the domain.

Also add a macro to disable those.

Given that these rules may be tried on every call (e.g. of `car`), it seems
valuable to assess whether disabling them provides good speed-ups without
introducing the need to enable (some of) them in too many places.