Added fast-cert mode for faster (but potentially unsound) book certification.
Quoting :DOC note-8-6:

  See [fast-cert] for a ``fast-cert'' mode for faster, but possibly
  unsound, book certification, in particular when using a saved
  executable that contains [local] events.  Thanks to Sol Swords for
  requesting such a capability and for helpful design discussions.

Thanks too to J for ideas for folding in future checks, including
perhaps ones considered by Sol.

In addition to running the usual make regression-everything (modulo
adding some comments), I have run a successful regression using
fast-cert mode on a near-final version, and I will do one on the final
version before pushing to github.  As discussed in the new Essay on
Fast-cert, I plan not to do future regressions using fast-cert mode.

.....
Some implementation level notes (most will want to ignore these!):
.....

There were many changes to support this work.  See :DOC fast-cert for
user-level information and see the Essay on Fast-cert for
implementation-level comments.  One change is that we now record in an
event-tuple whether the event is local (with new local-p argument to
make-event-tuple and other functions, and a new localp "field" of an
event-tuple).  The writing of *1* functions to expansion file has been
simplified (now done with function
write-*1*-defuns-to-expansion-file).  Eliminated unused state global
slow-apply$-action.  Deleted top-level-loop$-alist-rec and
top-level-loop$-alist (should have done so before when eliminating
*hcomp-loop$-alist* and *set-hcomp-loop$-alist*) and deleted
newly-defined-top-level-fns-rec and newly-defined-top-level-fns.

Modified debug printing for make-event based on make-event-debug so
that, as already documented, the depth is the expansion depth.  That
is the depth is incremented only during expansion, not when the
expansion is evaluated.

Made other small improvements.  For example, the bad-lisp-atomp error
message has been improved for the case of a hidden package.  Also, *1*
function definitions are now written from oldest to newest, rather
than in reverse order as was previously done.