Force recertification of regression (to fix build failure) and update *acl2-exports* with pending symbols.
An ACL2 structure for events (event-tuples) changed with the preceding
commit, so it is necessary to force recertification for regression
commands.  While I'm at it, I'm also adding three symbols into
*acl2-exports* whose addition was pending
(explain-giant-lambda-object, with-cbd, and set-fast-cert) to save
a separate need for mass recertification.

I'm checking this in before testing it to fix a build failure.  In the
very unlikely event that the *acl2-exports* change causes problems,
I'll fix that right away.