[PFCS] Improve formal semantics.
Extend proof trees to include, in a node for a relation constraint, the extended
assignment that must be in the assertions proved by all the subtrees. This
addition eliminates the need for a disjunction to take into account the case of
an empty list of subtrees, which is perhaps unusual because it corresponds to an
always true PFCS relation (with no constraints, that are taken conjunctively),
but allowed in both syntax and semantics. It also avoids the slight awkwardness
in picking a specific element out of a list of equal elements.